frontend: move forwarder under gateway
