+++ cmd/buildkitd/main.go
@@ -21,8 +21,9 @@ import (
 	registryremotecache "github.com/moby/buildkit/cache/remotecache/registry"
 	"github.com/moby/buildkit/control"
 	"github.com/moby/buildkit/frontend"
-	"github.com/moby/buildkit/frontend/dockerfile"
+	"github.com/moby/buildkit/frontend/dockerfile/builder"
 	"github.com/moby/buildkit/frontend/gateway"
+	"github.com/moby/buildkit/frontend/gateway/forwarder"
 	"github.com/moby/buildkit/session"
 	"github.com/moby/buildkit/solver/boltdbcachestorage"
 	"github.com/moby/buildkit/util/apicaps"
@@ -383,7 +384,7 @@ func newController(c *cli.Context, root string) (*control.Controller, error) {
 		return nil, err
 	}
 	frontends := map[string]frontend.Frontend{}
-	frontends["dockerfile.v0"] = dockerfile.NewDockerfileFrontend(wc)
+	frontends["dockerfile.v0"] = forwarder.NewGatewayForwarder(wc, builder.Build2)
 	frontends["gateway.v0"] = gateway.NewGatewayFrontend(wc)
 
 	cacheStorage, err := boltdbcachestorage.NewStore(filepath.Join(root, "cache.db"))
